[Fix]設定管理クラスの設定項目の名前空間化
- Settings,PedalSettings, SwitchBehavior, PedalModeを名前空間SettingsDefsに移動
- ioメソッドの追加
